refactoring of WeeklyProgress component styling
