--- src/components/Dashboard/OverallDashboard.module.css
@@ -1,94 +1,0 @@
-.Select {
-  width: 154px;
-  height: 33px;
-  flex-shrink: 0;
-  background-color: white;
-  border: 0;
-  border-radius: 4;
-  padding-left: 13px;
-  padding-right: 6px;
-  padding: auto;
-  padding-top: auto;
-}
-
-.SelectBox {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 100%;
-}
-
-.SelectTypography {
-  color: #8d8d8d;
-  font-family: var(--font-dm-sans);
-  font-size: 14px;
-  line-height: 18px;
-  letter-spacing: 0.42px;
-  margin-left: 16px;
-  margin-right: auto;
-}
-
-.SelectOptionTypography {
-  display: flex;
-  width: 153px;
-  height: 46px;
-  padding: 11px 24px;
-  justify-content: space-between;
-  align-items: center;
-  flex-shrink: 0;
-  align-self: stretch;
-  color: black;
-  text-align: center;
-  font-family: var(--font-poppins);
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 18px; /* 150% */
-  letter-spacing: 0.36px;
-  box-shadow: 12.02189px 14.59801px 34.34826px 3.43483px
-    rgba(112, 144, 176, 0.08);
-}
-
-.Chip {
-  width: fit-content;
-  padding: 0px 5px 0px 5px;
-  border-radius: 10px;
-  text-align: center;
-}
-
-.ChipTypography {
-  color: #2b3674;
-  font-family: var(--font-dm-sans);
-  font-weight: 700;
-  font-size: 10px;
-  line-height: 20.968px;
-  letter-spacing: -0.2px;
-}
-
-.DashboardTitleTypography {
-  font-family: var(--font-dm-sans);
-  font-weight: 700;
-  font-size: 29.705px;
-  font-style: normal;
-  line-height: 36.695px;
-  letter-spacing: -0.594px;
-}
-
-.ActiveIndicatorBox {
-  width: 95.232px;
-  height: 41.937px;
-  border-radius: 13.105px;
-  background-color: rgba(5, 205, 153, 0.1);
-  padding-top: 10.48px;
-  padding-bottom: 10.48px;
-  padding-left: 12px;
-}
-.ActiveIndicatorTypography {
-  color: #2b3674;
-  font-family: var(--font-dm-sans);
-  font-weight: 700;
-  font-size: 12.232px;
-  font-style: normal;
-  line-height: 20.968px;
-  letter-spacing: -0.245px;
-}
