+++ src/components/Dashboard/OverallDashboard/OverallDashboard.module.scss
@@ -1,0 +1,54 @@
+.Chip {
+  width: fit-content;
+  padding: 0px 5px 0px 5px;
+  border-radius: 10px;
+  text-align: center;
+}
+
+.ChipTypography {
+  color: #2b3674;
+  font-family: var(--font-dm-sans);
+  font-weight: 700;
+  font-size: 10px;
+  line-height: 20.968px;
+  letter-spacing: -0.2px;
+}
+
+.ActiveIndicatorBox {
+  width: 95.232px;
+  height: 41.937px;
+  border-radius: 13.105px;
+  background-color: rgba(5, 205, 153, 0.1);
+  padding-top: 10.48px;
+  padding-bottom: 10.48px;
+  padding-left: 12px;
+}
+.ActiveIndicatorTypography {
+  color: #2b3674;
+  font-family: var(--font-dm-sans);
+  font-weight: 700;
+  font-size: 12.232px;
+  font-style: normal;
+  line-height: 20.968px;
+  letter-spacing: -0.245px;
+}
+
+.OverallDashboardContainer {
+  height: 564px;
+  width: 100%;
+
+  .titleRow {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    margin-bottom: 51px;
+
+    .title {
+      font-family: var(--font-dm-sans);
+      font-weight: 700;
+      font-size: 29.705px;
+      font-style: normal;
+    }
+  }
+}
