+++ src/components/Graphs/index.ts
@@ -2,5 +2,6 @@ import LineChart from "./LineChart";
 import BarChart from "./BarChart";
 import StackedBarChart from "./StackedBarChart";
 import SmallDataBox from "./SmallDataBox";
+import WeeklyProgress from "./WeeklyProgress/WeeklyProgress";
 
-export { LineChart, BarChart, StackedBarChart, SmallDataBox };
+export { LineChart, BarChart, StackedBarChart, SmallDataBox, WeeklyProgress };
